requirements.txt
numpy
pandas
scipy
scikit-learn
librosa
sounddevice
joblib
matplotlib
seaborn
tqdm
pyserial
requests
